Change VolumeCornerIterator and move to hpp
