+++ tests/Unit/Domain/DomainTestHelpers.hpp
@@ -4,12 +4,17 @@
 #pragma once
 
 #include <boost/rational.hpp>
+#include <cstddef>
 #include <memory>
 #include <unordered_map>
 #include <unordered_set>
 #include <vector>
 
 #include "DataStructures/Tensor/IndexType.hpp"
+#include "Domain/Direction.hpp"
+#include "Utilities/ConstantExpressions.hpp"
+#include "Utilities/Gsl.hpp"
+#include "Utilities/MakeArray.hpp"
 
 template <size_t VolumeDim, typename TargetFrame>
 class Block;
@@ -23,6 +28,34 @@ template <size_t VolumeDim, typename TargetFrame>
 class Domain;
 template <size_t VolumeDim>
 class ElementId;
+
+// Iterates over the logical corners of a VolumeDim-dimensional cube.
+template <size_t VolumeDim>
+class VolumeCornerIterator {
+ public:
+  VolumeCornerIterator() noexcept = default;
+  explicit VolumeCornerIterator(size_t index) noexcept : index_(index) {}
+  void operator++() noexcept {
+    ++index_;
+    for (size_t i = 0; i < VolumeDim; i++) {
+      gsl::at(array_sides_, i) =
+          2 * get_nth_bit(index_, i) - 1 == 1 ? Side::Upper : Side::Lower;
+    }
+  }
+  explicit operator bool() const noexcept {
+    return index_ < two_to_the(VolumeDim);
+  }
+  const std::array<Side, VolumeDim>& operator()() const noexcept {
+    return array_sides_;
+  }
+  const std::array<Side, VolumeDim>& operator*() const noexcept {
+    return array_sides_;
+  }
+
+ private:
+  size_t index_ = 0;
+  std::array<Side, VolumeDim> array_sides_ = make_array<VolumeDim>(Side::Lower);
+};
 
 // Test that the Blocks in the Domain are constructed correctly.
 template <size_t VolumeDim>
